Se modifico la ruta de los archivos ya que se saco nalizador , analizador se le cambio de nombre ahora puede ser usador dede el priencipal
