--- Separador con lineas usando txt/Archivos de diccionario/operadores/operadores.txt
@@ -1,15 +1,0 @@
-<
->
-=
-+
--
-AND
-OR
-NOT
-:
-==
-!=
-<=
->=
-+=
--=
