--- Separador con lineas usando txt/Archivos de diccionario/palabras reservadas/palabrasReservadas.txt
@@ -1,10 +1,0 @@
-DEF
-IMPORT
-FOR
-WHILE
-TRUE
-FALSE
-IF
-ELSE
-ELIF
-RETURN
