--- Separador con lineas usando txt/Archivos de diccionario/patrones/patrones.txt
@@ -1,4 +1,0 @@
-^".*?"$,CADENA
-^\d+\.\d+$,NUMERO
-^\d+$,NUMERO
-^[a-zA-Z_][a-zA-Z0-9_]*$,IDENTIFICADOR
