+++ archivos/Archivos de diccionario/operadores/operadores.txt
