+++ archivos/Archivos de diccionario/palabras reservadas/palabrasReservadas.txt
