+++ archivos/Archivos de diccionario/patrones/patrones.txt
@@ -1,4 +1,4 @@
 ^".*?"$,CADENA
 ^\d+\.\d+$,NUMERO
 ^\d+$,NUMERO
-^[a-zA-Z_][a-zA-Z0-9_]*$,IDENTIFICADOR
+^[a-zA-Z_][a-zA-Z0-9_]*$,ID
